Add files via upload
It is a GUI tool built with a mixture of customtkinter and tkinter.ttk. It has the functions of adding to-do items, list display (with due date, status, note preview), multi-state management, double-click editing, graphical calendar selection of dates, detailed information viewing of notes with hyperlinks, Ctrl+S shortcut key saving, and export to Excel files. Supports using the Enter key to return to the new line in the note field, and Ctrl+Enter to save the comment.